notification-service/requirements.txt
brevo-python==3.0.7
confluent-kafka==2.3.0
python-dotenv==1.0.0